feat(akari): Save module data and process running time (#114)
* feat(akari): Add AkariDataModuleType

Save module information and processing times.

* feat(akari): Use AkariDataModuleType from the AkariRouter

Measure processing times and save them.

* feat(akari): Add logger options for the AkariRouter

Control whether to log information and duration.

* feat(akari): Log processing duration of each module

Calculate the time to take runnig modules.

* fix(akari): Prevent mutable default arguments

Generate new options if the argument is missing.

* fix(akari): Correct the docstring of the setModule

It was not the correct syntax to add documents.

* refactor(akari): Make the AkariDataModuleType as dataclass

It reduces the code amount and makes it easy to read.